requirements.txt
dash==2.0.0
dash-table==5.0.0
dash-bootstrap-components==1.0.1
flask==2.0.1
numba==0.54.1
numpy==1.19.2
pandas==1.1.3
plotly==5.3.1
pyyaml==5.4.1